refactor(gates): one walk over the licensed trees, not two
The license-header sweep and the env-var contract each carried their own
WalkDir over the same roots, with the same node_modules skip, the same
regular-file guard and the same generated-file filter.

Nothing was wrong with either, and that is the hazard: a duplicated walk drifts
one skip at a time, and the gate keeping the older copy quietly covers less
than the one that got the fix. Both already shared the ROOTS — a new Go tree
enrols in every gate at once — so what was duplicated was the loop around that
list, not the obligation.

walkHandWrittenGoFiles is the shared form, and it RETURNS THE COUNT rather than
just visiting. That is the reason it is a wrapper and not a filter each caller
repeats: a gate whose root yields nothing passes exactly like a clean one, and
only the caller knows whether an empty root is legitimate — extensions/ is,
backend/ is not. Handing back the number lets each decide without re-deriving
what "hand-written Go file" means.

Four sweeps route through it now: the header check and three in the env
contract that were repeating the same two-line filter. The one that also wants
non-test files keeps that condition, which is its own and not the shared walk's.

Both safeguards verified to survive the move — a root that does not exist still
fails on the walk error, and a count that stops incrementing still fails
mustHaveFiles.
